requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
selenium==4.1.0
sqlalchemy==1.4.29
pymysql==1.0.2
python-dotenv==0.19.2
numpy
pandas
scikit-learn==1.0.2
scipy==1.7.3
matplotlib==3.5.1
transformers==4.15.0
spacy==3.2.0
nltk==3.6.7
fasttext==0.9.2
lxml==4.7.1
